LPS-54881 move lucene specific QueryTranslator out of portal-service
