+++ portal-impl/src/com/liferay/portal/search/lucene/QueryTranslator.java
@@ -12,7 +12,10 @@
  * details.
  */
 
-package com.liferay.portal.kernel.search;
+package com.liferay.portal.search.lucene;
+
+import com.liferay.portal.kernel.search.ParseException;
+import com.liferay.portal.kernel.search.Query;
 
 /**
  * @author Raymond Augé
